Extract function for Test Framework rendering to ui-common
