+++ utbot-ui-commons/src/main/kotlin/org/utbot/intellij/plugin/ui/utils/DialogWindowUtils.kt
@@ -1,0 +1,20 @@
+package org.utbot.intellij.plugin.ui.utils
+
+import com.intellij.ui.ColoredListCellRenderer
+import com.intellij.ui.SimpleTextAttributes
+import org.utbot.framework.codegen.domain.TestFramework
+import javax.swing.JList
+
+fun createTestFrameworksRenderer(willBeInstalledLabel: String): ColoredListCellRenderer<TestFramework> {
+    return object : ColoredListCellRenderer<TestFramework>() {
+        override fun customizeCellRenderer(
+            list: JList<out TestFramework>, value: TestFramework,
+            index: Int, selected: Boolean, hasFocus: Boolean
+        ) {
+            this.append(value.displayName, SimpleTextAttributes.REGULAR_ATTRIBUTES)
+            if (!value.isInstalled) {
+                this.append(willBeInstalledLabel, SimpleTextAttributes.ERROR_ATTRIBUTES)
+            }
+        }
+    }
+}
